requirements.txt
pyyaml>=6.0
matplotlib>=3.7
numpy>=1.24
pytest>=7.0
requests>=2.31
beautifulsoup4>=4.12
